requirements.txt
aiohttp
aiosqlite
nonebot2
nonebot-adapter-onebot
loguru
